all kalman-filter classes now inherit a `BaseClassExtended` class, which provides common methods `to_dict`, `from_dict` and `__getitem__`
